+++ tests/unit/server/search_engine/test_elasticsearch_engine.py
@@ -101,7 +101,7 @@ async def test_banking_sentiment_dataset(elasticsearch_engine: ElasticSearchEngi
 
     await elasticsearch_engine.create_index(dataset)
 
-    await elasticsearch_engine.add_records(
+    await elasticsearch_engine.index_records(
         dataset,
         records=[
             await RecordFactory.create(
@@ -474,7 +474,7 @@ class TestSuiteElasticSearchEngine:
         records = sorted(dataset_for_pagination.records, key=lambda r: r.id)
         assert [record.id for record in records[offset : offset + limit]] == [item.record_id for item in results.items]
 
-    async def test_add_records(self, elasticsearch_engine: ElasticSearchEngine, opensearch: OpenSearch):
+    async def test_index_records(self, elasticsearch_engine: ElasticSearchEngine, opensearch: OpenSearch):
         text_fields = await TextFieldFactory.create_batch(5)
         dataset = await DatasetFactory.create(fields=text_fields, questions=[])
         records = await RecordFactory.create_batch(
@@ -487,7 +487,7 @@ class TestSuiteElasticSearchEngine:
         await _refresh_dataset(dataset)
 
         await elasticsearch_engine.create_index(dataset)
-        await elasticsearch_engine.add_records(dataset, records)
+        await elasticsearch_engine.index_records(dataset, records)
 
         index_name = index_name_for_dataset(dataset)
         opensearch.indices.refresh(index=index_name)
@@ -508,7 +508,7 @@ class TestSuiteElasticSearchEngine:
         await _refresh_dataset(dataset)
 
         await elasticsearch_engine.create_index(dataset)
-        await elasticsearch_engine.add_records(dataset, records)
+        await elasticsearch_engine.index_records(dataset, records)
 
         records_to_delete, records_to_keep = records[:5], records[5:]
         await elasticsearch_engine.delete_records(dataset, records_to_delete)